requirements.txt
flask
colorgram.py